[v0.9.8.1-qfit3-resol1-cln3-newf32-dbg2-sz10] size_t in gameblocks

+++ datatype.h
@@ -40,9 +40,9 @@ struct DATA {
 	ptrdiff_t	labels_end_idx;
 	ptrdiff_t	name	[MAXPLAYERS];
 
-	int 		gb_filled;
-	int 		gb_idx;
-	int			gb_allocated;
+	size_t 		gb_filled;
+	size_t 		gb_idx;
+	size_t		gb_allocated;
 
 	struct GAMEBLOCK *gb[MAXBLOCKS];
 };
